poc/requirements.txt
# POC Testing Requirements
ollama==0.1.6
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.0
easyocr==1.7.1
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
nvidia-ml-py3==7.352.0
requests==2.31.0
